+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.0.47
 
+- Remember recent search terms in the search screen
 - Add Instagram post embeds
 - Add setting to enable Instagram embeds, off by default
 - Make embedded tweet webview background transparent